Se avanza con el flujo de Pedidos. Tenemos BD completa y normalizada Clases de dominio creadas Clases de negocio creadas Acceso a datos con soporte de transacciones FrmPedido algunos combos ya trabajan correctamente.

+++ dominio/Pedido.cs
@@ -14,7 +14,7 @@ namespace dominio
 
      
        
-        public float PrecioTotal {  get; set; }
+        public decimal PrecioTotal {  get; set; }
         public DateTime Fecha { get; set; }
         public int IdCliente { get; set; }
         public Cliente Cliente { get; set; }
